backend/requirements.txt
bottleneck==1.3.5
flask-cors==4.0.0
flask-marshmallow==0.15.0
flask-migrate==4.0.5
marshmallow-sqlalchemy==0.29.0
# mkl-fft==1.3.6
# mkl-random==1.2.2
node==1.2.1
numexpr==2.8.4
pip-chill==1.0.3
python-dotenv==1.0.1
